Added custom exceptions for CourseService

+++ src/main/java/com/foxminded/services/CourseService.java
@@ -3,6 +3,8 @@ package com.foxminded.services;
 import com.foxminded.dao.CourseDao;
 import com.foxminded.dto.CourseDto;
 import com.foxminded.entities.Course;
+import com.foxminded.services.exceptions.EntityAlreadyExistsException;
+import com.foxminded.services.exceptions.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.PropertyMap;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,26 +38,46 @@ public class CourseService {
     }
 
     public CourseDto findById(int id) {
-        return mapper.map(dao.findById(id), CourseDto.class);
+        Course course = dao.findById(id);
+        if (course == null) {
+            throw new EntityNotFoundException("Not found course with id: " + id);
+        }
+        return mapper.map(course, CourseDto.class);
     }
 
     public CourseDto findByName(String name) {
-        return mapper.map(dao.findByName(name), CourseDto.class);
+        Course course = dao.findByName(name);
+        if (course == null) {
+            throw new EntityNotFoundException("Not found course with name: " + name);
+        }
+        return mapper.map(course, CourseDto.class);
     }
 
     public void add(CourseDto course) {
+        if (dao.findByName(course.getName()) != null) {
+            throw new EntityAlreadyExistsException("Course with name " + course.getName() + " already exists.");
+        }
         dao.add(mapper.map(course, Course.class));
     }
 
-    public void update(int id, CourseDto course) {
-        dao.update(id, mapper.map(course, Course.class));
+    public void update(int id, CourseDto courseDto) {
+        if (dao.findById(id) == null) {
+            throw new EntityNotFoundException("Not found course with id: " + id);
+        }
+        dao.update(id, mapper.map(courseDto, Course.class));
     }
 
     public void deleteById(int id) {
+        if (dao.findById(id) == null) {
+            throw new EntityNotFoundException("Not found course with id: " + id);
+        }
         dao.deleteById(id);
     }
 
     public void deleteByName(String name) {
+        if (dao.findByName(name) == null) {
+            throw new EntityNotFoundException("Not found course with name: " + name);
+        }
         dao.deleteByName(name);
     }
